refactor(core): rename tool-guards to tool-session
